refactor: use form components
