+++ components/login/LoginForm.tsx
@@ -1,7 +1,7 @@
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { FaUser, FaLock } from 'react-icons/fa6';
-import { Button, Form } from './LoginForm.style';
-import MainInput from '../ui/input/MainInput';
+import Button from '../form/Button';
+import Input from '../form/Input';
 
 interface FormInput {
   username: string;
@@ -14,26 +14,30 @@ function LoginForm() {
   const onSubmit: SubmitHandler<FormInput> = (data) => console.log(data);
 
   return (
-    <Form onSubmit={handleSubmit(onSubmit)}>
-      <MainInput
-        icon={<FaUser color="var(--primary1)" />}
-        name="username"
-        type="text"
-        placeholder="아이디"
-        register={register}
-        required
-      />
-      <MainInput
-        icon={<FaLock color="var(--primary1)" />}
-        name="password"
-        type="password"
-        placeholder="비밀번호"
-        register={register}
-        required
-      />
-      {/* <input type="checkbox" /> 로그인 유지 */}
-      <Button>로그인하기</Button>
-    </Form>
+    <form onSubmit={handleSubmit(onSubmit)}>
+      <div>
+        <Input
+          icon={<FaUser color="var(--primary1)" />}
+          name="username"
+          type="text"
+          placeholder="아이디"
+          register={register}
+          required
+        />
+        <Input
+          icon={<FaLock color="var(--primary1)" />}
+          name="password"
+          type="password"
+          placeholder="비밀번호"
+          register={register}
+          required
+        />
+      </div>
+      <div>
+        {/* <input type="checkbox" /> 로그인 유지 */}
+        <Button>로그인하기</Button>
+      </div>
+    </form>
   );
 }
